Remade in new structure.
