+++ @types/contracts.d.ts
@@ -2,6 +2,7 @@ export interface Config {
     Entry: string;
     Destination: string;
     Verbosity: Verbosity;
+    DedupePaths: string[];
 }
 export declare enum Verbosity {
     None = 0,
@@ -13,4 +14,5 @@ export interface ArgumentsValues {
     entry: string;
     dest: string;
     verbosity: Verbosity;
+    dedupePaths?: string[];
 }
