+++ src/contracts.ts
@@ -2,6 +2,7 @@ export interface Config {
     Entry: string;
     Destination: string;
     Verbosity: Verbosity;
+    DedupePaths: string[];
 }
 
 export enum Verbosity {
@@ -15,4 +16,5 @@ export interface ArgumentsValues {
     entry: string;
     dest: string;
     verbosity: Verbosity;
+    dedupePaths?: string[];
 }
